initial changes to retrieve acheron data from Drupal
